+++ src/mlpack/bindings/python/print_input_processing.hpp
@@ -464,13 +464,16 @@ void PrintInputProcessing(
 
   /** We want to generate code like the following:
    *
+   * cdef extern from "numpy/arrayobject.h":
+   *   void* PyArray_DATA(np.ndarray arr)
    * if param_name is not None:
    *   param_name_tuple = to_matrix_with_info(param_name)
    *   if len(param_name_tuple[0].shape) < 2:
    *     param_name_tuple[0].shape = (param_name_tuple[0].size,)
    *   param_name_mat = arma_numpy.numpy_to_matrix_d(param_name_tuple[0])
    *   SetParamWithInfo[mat](p, \<const string\> 'param_name',
-   *       dereference(param_name_mat), &param_name_tuple[1][0])
+   *       dereference(param_name_mat), 
+   *       \<const cbool*\> PyArray_DATA(param_name_dims))
    *   p.SetPassed(\<const string\> 'param_name')
    */
   std::cout << prefix << "cdef np.ndarray " << name << "_dims" << std::endl;
@@ -478,9 +481,11 @@ void PrintInputProcessing(
       << std::endl;
   if (!d.required)
   {
-    std::cout << prefix << "if " << name << " is not None:" << std::endl;
-    std::cout << prefix << "  " << name << "_tuple = to_matrix_with_info("
-        << name << ", dtype=np.double, copy=p.Has('copy_all_inputs'))"
+    std::cout << prefix << "cdef extern from \"numpy/arrayobject.h\":" << std::endl;
+    std::cout << prefix << "  void* PyArray_DATA(np.ndarray arr)" << std::endl;
+    std::cout << prefix << "if " << d.name << " is not None:" << std::endl;
+    std::cout << prefix << "  " << d.name << "_tuple = to_matrix_with_info("
+        << d.name << ", dtype=np.double, copy=p.Has('copy_all_inputs'))"
         << std::endl;
     std::cout << prefix << "  if len(" << name << "_tuple[0].shape"
         << ") < 2:" << std::endl;
@@ -491,15 +496,17 @@ void PrintInputProcessing(
     std::cout << prefix << "  " << name << "_dims = " << name
         << "_tuple[2]" << std::endl;
     std::cout << prefix << "  SetParamWithInfo[arma.Mat[double]](p, <const "
-        << "string> '" << d.name << "', dereference(" << name << "_mat), "
-        << "<const cbool*> " << name << "_dims.data)" << std::endl;
+        << "string> '" << d.name << "', dereference(" << d.name << "_mat), "
+        << "<const cbool*> PyArray_DATA(" << d.name << "_dims))" << std::endl;
     std::cout << prefix << "  p.SetPassed(<const string> '" << d.name
         << "')" << std::endl;
     std::cout << prefix << "  del " << name << "_mat" << std::endl;
   }
   else
   {
-    std::cout << prefix << name << "_tuple = to_matrix_with_info(" << name
+    std::cout << prefix << "cdef extern from \"numpy/arrayobject.h\":" << std::endl;
+    std::cout << prefix << "  void* PyArray_DATA(np.ndarray arr)" << std::endl;
+    std::cout << prefix << d.name << "_tuple = to_matrix_with_info(" << d.name
         << ", dtype=np.double, copy=p.Has('copy_all_inputs'))"
         << std::endl;
     std::cout << prefix << "if len(" << name << "_tuple[0].shape"
@@ -511,8 +518,8 @@ void PrintInputProcessing(
     std::cout << prefix << name << "_dims = " << name << "_tuple[2]"
         << std::endl;
     std::cout << prefix << "SetParamWithInfo[arma.Mat[double]](p, <const "
-        << "string> '" << d.name << "', dereference(" << name << "_mat), "
-        << "<const cbool*> " << name << "_dims.data)" << std::endl;
+        << "string> '" << d.name << "', dereference(" << d.name << "_mat), "
+        << "<const cbool*> PyArray_DATA(" << d.name << "_dims))" << std::endl;
     std::cout << prefix << "p.SetPassed(<const string> '" << d.name << "')"
         << std::endl;
     std::cout << prefix << "del " << name << "_mat" << std::endl;